refactor: split recovery/engine_test.go into focused test files
Split the 1286-line monolith into 5 files by responsibility:
- failure_test.go: FailureType classification and string/recoverable
- cache_test.go: IntentCache store, lookup, TTL, LRU, invalidation
- engine_recovery_test.go: RecoveryEngine unit tests (mock matcher)
- engine_scenario_test.go: realistic end-to-end recovery scenarios
- config_test.go: DefaultRecoveryConfig validation

Removed section demarcation comments (========) that were
signaling multiple responsibilities in a single file.

All 167 tests pass with -race.
